Add module for creating certificates for tests
Motivation:
You need a certificate in order to test much of anything TLS related.
This is true for both Netty, and anyone else using Netty.
The SelfSignedCertificate was fine for just Netty usage, but is reaching the limits of its API when others are using it.
It relies on internal JDK classes (which have since been restricted), or it relies on BouncyCastle being available

Modification:
Add a separate module for creating test certificates, and implement the certificate signing from scratch such that neither internal JDK classes, nor BouncyCastle, is required.

Result:
Netty and her users can now reliably create test certificates, without relying on particular JDK or BouncyCastle versions.